LPD-32101 Move to core upgrade and adapt naming
